Split test files
