+++ src/tests/api/test_orders.py
@@ -25,10 +25,8 @@ import json
 from decimal import Decimal
 from unittest import mock
 
-import freezegun
 import pytest
 from django.core import mail as djmail
-from django.core.files.base import ContentFile
 from django.utils.timezone import now
 from django_countries.fields import Country
 from django_scopes import scopes_disabled
@@ -36,15 +34,8 @@ from pytz import UTC
 from stripe.error import APIConnectionError
 from tests.plugins.stripe.test_provider import MockedCharge
 
-from pretix.base.models import (
-    InvoiceAddress, Order, OrderPosition, Question, SeatingPlan,
-)
-from pretix.base.models.orders import (
-    CartPosition, OrderFee, OrderPayment, OrderRefund, QuestionAnswer,
-)
-from pretix.base.services.invoices import (
-    generate_cancellation, generate_invoice,
-)
+from pretix.base.models import InvoiceAddress, Order, OrderPosition
+from pretix.base.models.orders import OrderFee, OrderPayment, OrderRefund
 
 
 @pytest.fixture
@@ -980,179 +971,6 @@ def test_orderposition_delete(token_client, organizer, event, order, item, quest
     assert order.total == Decimal('23.25')
 
 
-@pytest.fixture
-def invoice(order):
-    testtime = datetime.datetime(2017, 12, 10, 10, 0, 0, tzinfo=UTC)
-
-    with mock.patch('django.utils.timezone.now') as mock_now:
-        mock_now.return_value = testtime
-        return generate_invoice(order)
-
-
-TEST_INVOICE_RES = {
-    "order": "FOO",
-    "number": "DUMMY-00001",
-    "is_cancellation": False,
-    "invoice_from_name": "",
-    "invoice_from": "",
-    "invoice_from_zipcode": "",
-    "invoice_from_city": "",
-    "invoice_from_country": None,
-    "invoice_from_tax_id": "",
-    "invoice_from_vat_id": "",
-    "invoice_to": "Sample company\nNew Zealand\nVAT-ID: DE123",
-    "invoice_to_company": "Sample company",
-    "invoice_to_name": "",
-    "invoice_to_street": "",
-    "invoice_to_zipcode": "",
-    "invoice_to_city": "",
-    "invoice_to_state": "",
-    "invoice_to_country": "NZ",
-    "invoice_to_vat_id": "DE123",
-    "invoice_to_beneficiary": "",
-    "custom_field": None,
-    "date": "2017-12-10",
-    "refers": None,
-    "locale": "en",
-    "introductory_text": "",
-    "internal_reference": "",
-    "additional_text": "",
-    "payment_provider_text": "",
-    "footer_text": "",
-    "foreign_currency_display": None,
-    "foreign_currency_rate": None,
-    "foreign_currency_rate_date": None,
-    "lines": [
-        {
-            "position": 1,
-            "description": "Budget Ticket<br />Attendee: Peter",
-            'subevent': None,
-            'event_date_from': '2017-12-27T10:00:00Z',
-            'event_date_to': None,
-            'event_location': None,
-            'attendee_name': 'Peter',
-            'item': None,
-            'variation': None,
-            'fee_type': None,
-            'fee_internal_type': None,
-            "gross_value": "23.00",
-            "tax_value": "0.00",
-            "tax_name": "",
-            "tax_rate": "0.00"
-        },
-        {
-            "position": 2,
-            "description": "Payment fee",
-            'subevent': None,
-            'event_date_from': '2017-12-27T10:00:00Z',
-            'event_date_to': None,
-            'event_location': None,
-            'attendee_name': None,
-            'fee_type': "payment",
-            'fee_internal_type': None,
-            'item': None,
-            'variation': None,
-            "gross_value": "0.25",
-            "tax_value": "0.05",
-            "tax_name": "",
-            "tax_rate": "19.00"
-        }
-    ]
-}
-
-
-@pytest.mark.django_db
-def test_invoice_list(token_client, organizer, event, order, item, invoice):
-    res = dict(TEST_INVOICE_RES)
-    res['lines'][0]['item'] = item.pk
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/'.format(organizer.slug, event.slug))
-    assert resp.status_code == 200
-    assert [res] == resp.data['results']
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?order=FOO'.format(organizer.slug, event.slug))
-    assert [res] == resp.data['results']
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?order=BAR'.format(organizer.slug, event.slug))
-    assert [] == resp.data['results']
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?number={}'.format(
-        organizer.slug, event.slug, invoice.number))
-    assert [res] == resp.data['results']
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?number=XXX'.format(
-        organizer.slug, event.slug))
-    assert [] == resp.data['results']
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?locale=en'.format(
-        organizer.slug, event.slug))
-    assert [res] == resp.data['results']
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?locale=de'.format(
-        organizer.slug, event.slug))
-    assert [] == resp.data['results']
-
-    with scopes_disabled():
-        ic = generate_cancellation(invoice)
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?is_cancellation=false'.format(
-        organizer.slug, event.slug))
-    assert [res] == resp.data['results']
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?is_cancellation=true'.format(
-        organizer.slug, event.slug))
-    assert len(resp.data['results']) == 1
-    assert resp.data['results'][0]['number'] == ic.number
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?refers={}'.format(
-        organizer.slug, event.slug, invoice.number))
-    assert len(resp.data['results']) == 1
-    assert resp.data['results'][0]['number'] == ic.number
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/?refers={}'.format(
-        organizer.slug, event.slug, ic.number))
-    assert [] == resp.data['results']
-
-
-@pytest.mark.django_db
-def test_invoice_detail(token_client, organizer, event, item, invoice):
-    res = dict(TEST_INVOICE_RES)
-    res['lines'][0]['item'] = item.pk
-
-    resp = token_client.get('/api/v1/organizers/{}/events/{}/invoices/{}/'.format(organizer.slug, event.slug,
-                                                                                  invoice.number))
-    assert resp.status_code == 200
-    assert res == resp.data
-
-
-@pytest.mark.django_db
-def test_invoice_regenerate(token_client, organizer, event, invoice):
-    organizer.settings.invoice_regenerate_allowed = True
-    with scopes_disabled():
-        InvoiceAddress.objects.filter(order=invoice.order).update(company="ACME Ltd")
-
-    with freezegun.freeze_time("2017-12-10"):
-        resp = token_client.post('/api/v1/organizers/{}/events/{}/invoices/{}/regenerate/'.format(
-            organizer.slug, event.slug, invoice.number
-        ))
-        assert resp.status_code == 204
-        invoice.refresh_from_db()
-        assert "ACME Ltd" in invoice.invoice_to
-
-
-@pytest.mark.django_db
-def test_invoice_reissue(token_client, organizer, event, invoice):
-    with scopes_disabled():
-        InvoiceAddress.objects.filter(order=invoice.order).update(company="ACME Ltd")
-
-    resp = token_client.post('/api/v1/organizers/{}/events/{}/invoices/{}/reissue/'.format(
-        organizer.slug, event.slug, invoice.number
-    ))
-    assert resp.status_code == 204
-    invoice.refresh_from_db()
-    assert "ACME Ltd" not in invoice.invoice_to
-    with scopes_disabled():
-        assert invoice.order.invoices.count() == 3
-        invoice = invoice.order.invoices.last()
-    assert "ACME Ltd" in invoice.invoice_to
-
-
 @pytest.mark.django_db
 def test_order_mark_paid_pending(token_client, organizer, event, order):
     resp = token_client.post(
@@ -1630,2385 +1448,6 @@ def test_order_invalid_state_deny(token_client, organizer, event, order):
     assert resp.status_code == 400
 
 
-ORDER_CREATE_PAYLOAD = {
-    "email": "dummy@dummy.test",
-    "phone": "+49622112345",
-    "locale": "en",
-    "sales_channel": "web",
-    "fees": [
-        {
-            "fee_type": "payment",
-            "value": "0.25",
-            "description": "",
-            "internal_type": "",
-            "tax_rule": None
-        }
-    ],
-    "payment_provider": "banktransfer",
-    "invoice_address": {
-        "is_business": False,
-        "company": "Sample company",
-        "name_parts": {"full_name": "Fo"},
-        "street": "Bar",
-        "state": "",
-        "zipcode": "",
-        "city": "Sample City",
-        "country": "NZ",
-        "internal_reference": "",
-        "vat_id": ""
-    },
-    "positions": [
-        {
-            "positionid": 1,
-            "item": 1,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "company": "FOOCORP",
-            "answers": [
-                {
-                    "question": 1,
-                    "answer": "S",
-                    "options": []
-                }
-            ],
-            "subevent": None
-        }
-    ],
-}
-
-
-@pytest.mark.django_db
-def test_order_create(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    with scopes_disabled():
-        customer = organizer.customers.create()
-    res['customer'] = customer.identifier
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-    assert o.customer == customer
-    assert o.email == "dummy@dummy.test"
-    assert o.phone == "+49622112345"
-    assert o.locale == "en"
-    assert o.total == Decimal('23.25')
-    assert o.status == Order.STATUS_PENDING
-    assert o.sales_channel == "web"
-    assert not o.testmode
-
-    with scopes_disabled():
-        p = o.payments.first()
-    assert p.provider == "banktransfer"
-    assert p.amount == o.total
-    assert p.state == "created"
-
-    with scopes_disabled():
-        fee = o.fees.first()
-    assert fee.fee_type == "payment"
-    assert fee.value == Decimal('0.25')
-    ia = o.invoice_address
-    assert ia.company == "Sample company"
-    assert ia.name_parts == {"full_name": "Fo", "_scheme": "full"}
-    assert ia.name_cached == "Fo"
-    with scopes_disabled():
-        assert o.positions.count() == 1
-        pos = o.positions.first()
-    assert pos.item == item
-    assert pos.price == Decimal("23.00")
-    assert pos.attendee_name_parts == {"full_name": "Peter", "_scheme": "full"}
-    assert pos.company == "FOOCORP"
-    with scopes_disabled():
-        answ = pos.answers.first()
-    assert answ.question == question
-    assert answ.answer == "S"
-    with scopes_disabled():
-        assert o.transactions.count() == 2
-
-
-@pytest.mark.django_db
-def test_order_create_simulate(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    question.type = Question.TYPE_CHOICE_MULTIPLE
-    question.save()
-    with scopes_disabled():
-        opt = question.options.create(answer="L")
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['answers'][0]['options'] = [opt.pk]
-    res['simulate'] = True
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert Order.objects.count() == 0
-        assert QuestionAnswer.objects.count() == 0
-        assert OrderPosition.objects.count() == 0
-        assert OrderFee.objects.count() == 0
-        assert InvoiceAddress.objects.count() == 0
-    d = resp.data
-    del d['last_modified']
-    del d['secret']
-    del d['url']
-    del d['expires']
-    del d['invoice_address']['last_modified']
-    del d['positions'][0]['secret']
-    assert d == {
-        'code': 'PREVIEW',
-        'status': 'n',
-        'testmode': False,
-        'email': 'dummy@dummy.test',
-        'phone': '+49622112345',
-        'customer': None,
-        'locale': 'en',
-        'datetime': None,
-        'payment_date': None,
-        'payment_provider': None,
-        'fees': [
-            {
-                'id': 0,
-                'fee_type': 'payment',
-                'value': '0.25',
-                'description': '',
-                'internal_type': '',
-                'tax_rate': '0.00',
-                'tax_value': '0.00',
-                'tax_rule': None,
-                'canceled': False
-            }
-        ],
-        'total': '23.25',
-        'comment': '',
-        "custom_followup_at": None,
-        'invoice_address': {
-            'is_business': False,
-            'company': 'Sample company',
-            'name': 'Fo',
-            'name_parts': {'full_name': 'Fo', '_scheme': 'full'},
-            'street': 'Bar',
-            'zipcode': '',
-            'city': 'Sample City',
-            'country': 'NZ',
-            'state': '',
-            'vat_id': '',
-            'vat_id_validated': False,
-            'internal_reference': ''
-        },
-        'positions': [
-            {
-                'id': 0,
-                'order': '',
-                'positionid': 1,
-                'item': item.pk,
-                'variation': None,
-                'price': '23.00',
-                'attendee_name': 'Peter',
-                'attendee_name_parts': {'full_name': 'Peter', '_scheme': 'full'},
-                'attendee_email': None,
-                'voucher': None,
-                'tax_rate': '0.00',
-                'tax_value': '0.00',
-                'addon_to': None,
-                'subevent': None,
-                'checkins': [],
-                'downloads': [],
-                'answers': [
-                    {'question': question.pk, 'answer': 'L', 'question_identifier': 'ABC',
-                     'options': [opt.pk],
-                     'option_identifiers': [opt.identifier]}
-                ],
-                'tax_rule': None,
-                'pseudonymization_id': 'PREVIEW',
-                'seat': None,
-                'company': "FOOCORP",
-                'street': None,
-                'city': None,
-                'zipcode': None,
-                'state': None,
-                'country': None,
-                'canceled': False
-            }
-        ],
-        'downloads': [],
-        'checkin_attention': False,
-        'payments': [],
-        'refunds': [],
-        'require_approval': False,
-        'sales_channel': 'web',
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_positionids_addons_simulated(token_client, organizer, event, item, quota):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "positionid": 2,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": 1,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    res['simulate'] = True
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    del resp.data['positions'][0]['secret']
-    del resp.data['positions'][1]['secret']
-    assert [dict(f) for f in resp.data['positions']] == [
-        {'id': 0, 'order': '', 'positionid': 1, 'item': item.pk, 'variation': None, 'price': '23.00',
-         'attendee_name': 'Peter', 'attendee_name_parts': {'full_name': 'Peter', '_scheme': 'full'}, 'company': None,
-         'street': None, 'zipcode': None, 'city': None, 'country': None, 'state': None, 'attendee_email': None,
-         'voucher': None, 'tax_rate': '0.00', 'tax_value': '0.00',
-         'addon_to': None, 'subevent': None, 'checkins': [], 'downloads': [], 'answers': [], 'tax_rule': None,
-         'pseudonymization_id': 'PREVIEW', 'seat': None, 'canceled': False},
-        {'id': 0, 'order': '', 'positionid': 2, 'item': item.pk, 'variation': None, 'price': '23.00',
-         'attendee_name': 'Peter', 'attendee_name_parts': {'full_name': 'Peter', '_scheme': 'full'}, 'company': None,
-         'street': None, 'zipcode': None, 'city': None, 'country': None, 'state': None, 'attendee_email': None,
-         'voucher': None, 'tax_rate': '0.00', 'tax_value': '0.00',
-         'addon_to': 1, 'subevent': None, 'checkins': [], 'downloads': [], 'answers': [], 'tax_rule': None,
-         'pseudonymization_id': 'PREVIEW', 'seat': None, 'canceled': False}
-    ]
-
-
-@pytest.mark.django_db
-def test_order_create_autocheckin(token_client, organizer, event, item, quota, question, clist_autocheckin):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert "web" in clist_autocheckin.auto_checkin_sales_channels
-        assert o.positions.first().checkins.first().auto_checked_in
-
-    clist_autocheckin.auto_checkin_sales_channels = []
-    clist_autocheckin.save()
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert clist_autocheckin.auto_checkin_sales_channels == []
-        assert o.positions.first().checkins.count() == 0
-
-
-@pytest.mark.django_db
-def test_order_create_require_approval(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['require_approval'] = True
-    res['send_email'] = True
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.require_approval
-    assert len(djmail.outbox) == 1
-    assert djmail.outbox[0].subject == "Your order: {}".format(resp.data['code'])
-    assert "approval" in djmail.outbox[0].body
-
-
-@pytest.mark.django_db
-def test_order_create_invoice_address_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['invoice_address']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        with pytest.raises(InvoiceAddress.DoesNotExist):
-            o.invoice_address
-
-
-@pytest.mark.django_db
-def test_order_create_sales_channel_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['sales_channel']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-    assert o.sales_channel == "web"
-
-
-@pytest.mark.django_db
-def test_order_create_sales_channel_invalid(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['sales_channel'] = 'foo'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'sales_channel': ['Unknown sales channel.']}
-
-
-@pytest.mark.django_db
-def test_order_create_in_test_mode(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['testmode'] = True
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-    assert o.testmode
-
-
-@pytest.mark.django_db
-def test_order_create_in_test_mode_saleschannel_limited(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['testmode'] = True
-    res['sales_channel'] = 'baz'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'testmode': ['This sales channel does not provide support for test mode.']}
-
-
-@pytest.mark.django_db
-def test_order_create_attendee_name_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['attendee_name'] = None
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['attendee_name_parts']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.positions.first().attendee_name_parts == {}
-
-
-@pytest.mark.django_db
-def test_order_create_legacy_attendee_name(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['attendee_name'] = 'Peter'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    del res['positions'][0]['attendee_name_parts']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.positions.first().attendee_name_parts == {"_legacy": "Peter"}
-
-
-@pytest.mark.django_db
-def test_order_create_legacy_invoice_name(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['invoice_address']['name'] = 'Peter'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    del res['invoice_address']['name_parts']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.invoice_address.name_parts == {"_legacy": "Peter"}
-
-
-@pytest.mark.django_db
-def test_order_create_code_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['code'] = 'ABCDE'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-    assert o.code == "ABCDE"
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'code': ['This order code is already in use.']}
-
-    res['code'] = 'ABaDE'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'code': ['This order code contains invalid characters.']}
-
-
-@pytest.mark.django_db
-def test_order_email_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['email']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-    assert not o.email
-
-
-@pytest.mark.django_db
-def test_order_create_payment_provider_optional_free(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['price'] = '0.00'
-    res['positions'][0]['status'] = 'p'
-    del res['payment_provider']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert not o.payments.exists()
-
-
-@pytest.mark.django_db
-def test_order_create_payment_info_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-    res['payment_info'] = {
-        'foo': {
-            'bar': [1, 2],
-            'test': False
-        }
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-
-        p = o.payments.first()
-    assert p.provider == "banktransfer"
-    assert p.amount == o.total
-    assert json.loads(p.info) == res['payment_info']
-
-
-@pytest.mark.django_db
-def test_order_create_position_secret_optional(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.positions.first().secret
-
-    res['positions'][0]['secret'] = "aaa"
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.positions.first().secret == "aaa"
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-
-    assert resp.data == {'positions': [{'secret': ['You cannot assign a position secret that already exists.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_tax_rules(token_client, organizer, event, item, quota, question, taxrule):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'][0]['tax_rule'] = taxrule.pk
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    item.tax_rule = taxrule
-    item.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        fee = o.fees.first()
-    assert fee.fee_type == "payment"
-    assert fee.value == Decimal('0.25')
-    assert fee.tax_rate == Decimal('19.00')
-    assert fee.tax_rule == taxrule
-    ia = o.invoice_address
-    assert ia.company == "Sample company"
-    with scopes_disabled():
-        pos = o.positions.first()
-    assert pos.item == item
-    assert pos.tax_rate == Decimal('19.00')
-    assert pos.tax_value == Decimal('3.67')
-    assert pos.tax_rule == taxrule
-
-
-@pytest.mark.django_db
-def test_order_create_fee_type_validation(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'][0]['fee_type'] = 'unknown'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'fees': [{'fee_type': ['"unknown" is not a valid choice.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_fee_as_percentage(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'][0]['_treat_value_as_percentage'] = True
-    res['fees'][0]['value'] = '10.00'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        fee = o.fees.first()
-        assert fee.value == Decimal('2.30')
-        assert o.total == Decimal('25.30')
-
-
-@pytest.mark.django_db
-def test_order_create_fee_with_auto_tax(token_client, organizer, event, item, quota, question, taxrule):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'][0]['_split_taxes_like_products'] = True
-    res['fees'][0]['_treat_value_as_percentage'] = True
-    res['fees'][0]['value'] = '10.00'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    item.tax_rule = taxrule
-    item.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        fee = o.fees.first()
-        assert fee.value == Decimal('2.30')
-        assert fee.tax_rate == Decimal('19.00')
-        assert o.total == Decimal('25.30')
-
-
-@pytest.mark.django_db
-def test_order_create_negative_fee_with_auto_tax(token_client, organizer, event, item, quota, question, taxrule):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'][0]['_split_taxes_like_products'] = True
-    res['fees'][0]['value'] = '-10.00'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    item.tax_rule = taxrule
-    item.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        fee = o.fees.first()
-        assert fee.value == Decimal('-10.00')
-        assert fee.tax_value == Decimal('-1.60')
-        assert fee.tax_rate == Decimal('19.00')
-        assert o.total == Decimal('13.00')
-
-
-@pytest.mark.django_db
-def test_order_create_tax_rule_wrong_event(token_client, organizer, event, item, quota, question, taxrule2):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'][0]['tax_rule'] = taxrule2.pk
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'fees': [{'tax_rule': ['The specified tax rate does not belong to this event.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_subevent_not_allowed(token_client, organizer, event, item, quota, question, subevent2):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['subevent'] = subevent2.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'subevent': ['You cannot set a subevent for this event.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_empty(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'] = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': ['An order cannot be empty.']}
-
-
-@pytest.mark.django_db
-def test_order_create_subevent_validation(token_client, organizer, event, item, subevent, subevent2, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'subevent': ['You need to set a subevent.']}]}
-
-    res['positions'][0]['subevent'] = subevent2.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'subevent': ['The specified subevent does not belong to this event.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_item_validation(token_client, organizer, event, item, item2, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    item.active = False
-    item.save()
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'item': ['The specified item is not active.']}]}
-    item.active = True
-    item.save()
-
-    res['positions'][0]['item'] = item2.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'item': ['The specified item does not belong to this event.']}]}
-
-    with scopes_disabled():
-        var2 = item2.variations.create(value="A")
-        quota.variations.add(var2)
-
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['variation'] = var2.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'variation': ['You cannot specify a variation for this item.']}]}
-
-    with scopes_disabled():
-        var1 = item.variations.create(value="A")
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['variation'] = var1.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'item': ['The product "Budget Ticket" is not assigned to a quota.']}]}
-
-    with scopes_disabled():
-        quota.variations.add(var1)
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-    res['positions'][0]['variation'] = var2.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [{'variation': ['The specified variation does not belong to the specified item.']}]}
-
-    res['positions'][0]['variation'] = None
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'variation': ['You should specify a variation for this item.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_subevent_disabled(token_client, organizer, event, item, subevent, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['subevent'] = subevent.pk
-    s = item.subeventitem_set.create(subevent=subevent, disabled=True)
-    quota.subevent = subevent
-    quota.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'item': ['The product "Budget Ticket" is not available on this date.']}]}
-
-    s.delete()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-
-@pytest.mark.django_db
-def test_order_create_subevent_variation_disabled(token_client, organizer, event, item, subevent, quota, question):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        var = item2.variations.create(default_price=12, value="XS")
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item2.pk
-    res['positions'][0]['variation'] = var.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['subevent'] = subevent.pk
-    s = var.subeventitemvariation_set.create(subevent=subevent, disabled=True)
-    quota.subevent = subevent
-    quota.items.add(item2)
-    quota.variations.add(var)
-    quota.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'item': ['The product "Budget Ticket" is not available on this date.']}]}
-
-    s.delete()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-
-@pytest.mark.django_db
-def test_order_create_positionids_addons(token_client, organizer, event, item, quota):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "positionid": 2,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": 1,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos1 = o.positions.first()
-        pos2 = o.positions.last()
-    assert pos2.addon_to == pos1
-
-
-@pytest.mark.django_db
-def test_order_create_positionid_validation(token_client, organizer, event, item, quota):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "positionid": 2,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": 2,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {},
-            {
-                'addon_to': [
-                    'If you set addon_to, you need to make sure that the '
-                    'referenced position ID exists and is transmitted directly '
-                    'before its add-ons.'
-                ]
-            }
-        ]
-    }
-
-    res['positions'] = [
-        {
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": 2,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [
-        {'positionid': ["If you set addon_to on any position, you need to specify position IDs manually."]},
-        {'positionid': ["If you set addon_to on any position, you need to specify position IDs manually."]}
-    ]}
-
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {},
-            {
-                'positionid': ['If you set position IDs manually, you need to do so for all positions.']
-            }
-        ]
-    }
-
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "positionid": 3,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {},
-            {
-                'positionid': ['Position IDs need to be consecutive.']
-            }
-        ]
-    }
-
-    res['positions'] = [
-        {
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.positions.first().positionid == 1
-        assert o.positions.last().positionid == 2
-
-
-@pytest.mark.django_db
-def test_order_create_answer_validation(token_client, organizer, event, item, quota, question, question2):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question2.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [{'answers': [{'question': ['The specified question does not belong to this event.']}]}]}
-
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['answers'][0]['options'] = [question.options.first().pk]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'answers': [
-        {'non_field_errors': ['You should not specify options if the question is not of a choice type.']}]}]}
-
-    question.type = Question.TYPE_CHOICE
-    question.save()
-    res['positions'][0]['answers'][0]['options'] = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [
-        {'answers': [{'non_field_errors': ['You need to specify options if the question is of a choice type.']}]}]}
-
-    with scopes_disabled():
-        question2.options.create(answer="L")
-    with scopes_disabled():
-        res['positions'][0]['answers'][0]['options'] = [
-            question2.options.first().pk,
-        ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [{'answers': [{'non_field_errors': ['The specified option does not belong to this question.']}]}]}
-
-    with scopes_disabled():
-        question.options.create(answer="L")
-    with scopes_disabled():
-        res['positions'][0]['answers'][0]['options'] = [
-            question.options.first().pk,
-            question.options.last().pk,
-        ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [{'answers': [{'non_field_errors': ['You can specify at most one option for this question.']}]}]}
-
-    r = token_client.post(
-        '/api/v1/upload',
-        data={
-            'media_type': 'image/png',
-            'file': ContentFile('file.png', 'invalid png content')
-        },
-        format='upload',
-        HTTP_CONTENT_DISPOSITION='attachment; filename="file.png"',
-    )
-    assert r.status_code == 201
-    file_id_png = r.data['id']
-    res['positions'][0]['answers'][0]['options'] = []
-    res['positions'][0]['answers'][0]['answer'] = file_id_png
-    question.type = Question.TYPE_FILE
-    question.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.file
-    assert answ.answer.startswith("file://")
-
-    question.type = Question.TYPE_CHOICE_MULTIPLE
-    question.save()
-    with scopes_disabled():
-        res['positions'][0]['answers'][0]['options'] = [
-            question.options.first().pk,
-            question.options.last().pk,
-        ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.question == question
-    assert answ.answer == "XL, L"
-
-    question.type = Question.TYPE_NUMBER
-    question.save()
-    res['positions'][0]['answers'][0]['options'] = []
-    res['positions'][0]['answers'][0]['answer'] = '3.45'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.answer == "3.45"
-
-    question.type = Question.TYPE_NUMBER
-    question.save()
-    res['positions'][0]['answers'][0]['options'] = []
-    res['positions'][0]['answers'][0]['answer'] = 'foo'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'answers': [{'non_field_errors': ['A valid number is required.']}]}]}
-
-    question.type = Question.TYPE_BOOLEAN
-    question.save()
-    res['positions'][0]['answers'][0]['options'] = []
-    res['positions'][0]['answers'][0]['answer'] = 'True'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.answer == "True"
-
-    question.type = Question.TYPE_BOOLEAN
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = '0'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.answer == "False"
-
-    question.type = Question.TYPE_BOOLEAN
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = 'bla'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [{'answers': [{'non_field_errors': ['Please specify "true" or "false" for boolean questions.']}]}]}
-
-    question.type = Question.TYPE_DATE
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = '2018-05-14'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.answer == "2018-05-14"
-
-    question.type = Question.TYPE_DATE
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = 'bla'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'answers': [
-        {'non_field_errors': ['Date has wrong format. Use one of these formats instead: YYYY-MM-DD.']}]}]}
-
-    question.type = Question.TYPE_DATETIME
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = '2018-05-14T13:00:00Z'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.answer == "2018-05-14 13:00:00+00:00"
-
-    question.type = Question.TYPE_DATETIME
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = 'bla'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'answers': [{'non_field_errors': [
-        'Datetime has wrong format. Use one of these formats instead: '
-        'YYYY-MM-DDThh:mm[:ss[.uuuuuu]][+HH:MM|-HH:MM|Z].']}]}]}
-
-    question.type = Question.TYPE_TIME
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = '13:00:00'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        pos = o.positions.first()
-        answ = pos.answers.first()
-    assert answ.answer == "13:00:00"
-
-    question.type = Question.TYPE_TIME
-    question.save()
-    res['positions'][0]['answers'][0]['answer'] = 'bla'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'answers': [
-        {'non_field_errors': ['Time has wrong format. Use one of these formats instead: hh:mm[:ss[.uuuuuu]].']}]}]}
-
-
-@pytest.mark.django_db
-def test_order_create_quota_validation(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "answers": [],
-            "subevent": None
-        },
-        {
-            "positionid": 2,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": 1,
-            "answers": [],
-            "subevent": None
-        }
-    ]
-
-    quota.size = 0
-    quota.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'item': ['There is not enough quota available on quota "Budget Quota" to perform the operation.']},
-            {'item': ['There is not enough quota available on quota "Budget Quota" to perform the operation.']},
-        ]
-    }
-
-    quota.size = 1
-    quota.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {},
-            {'item': ['There is not enough quota available on quota "Budget Quota" to perform the operation.']},
-        ]
-    }
-
-    res['force'] = True
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-
-@pytest.mark.django_db
-def test_order_create_quota_consume_cart(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-
-    with scopes_disabled():
-        cr = CartPosition.objects.create(
-            event=event, cart_id="uxLJBUMEcnxOLI2EuxLYN1hWJq9GKu4yWL9FEgs2m7M0vdFi@api", item=item,
-            price=23,
-            expires=now() + datetime.timedelta(hours=3)
-        )
-
-    quota.size = 1
-    quota.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'item': ['There is not enough quota available on quota "Budget Quota" to perform the operation.']},
-        ]
-    }
-
-    res['consume_carts'] = [cr.cart_id]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert not CartPosition.objects.filter(pk=cr.pk).exists()
-
-
-@pytest.mark.django_db
-def test_order_create_quota_consume_cart_expired(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-
-    with scopes_disabled():
-        cr = CartPosition.objects.create(
-            event=event, cart_id="uxLJBUMEcnxOLI2EuxLYN1hWJq9GKu4yWL9FEgs2m7M0vdFi@api", item=item,
-            price=23,
-            expires=now() - datetime.timedelta(hours=3)
-        )
-
-    quota.size = 0
-    quota.save()
-    res['consume_carts'] = [cr.cart_id]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'item': ['There is not enough quota available on quota "Budget Quota" to perform the operation.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_free(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['fees'] = []
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['price'] = '0.00'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-    assert o.total == Decimal('0.00')
-    assert o.status == Order.STATUS_PAID
-
-    with scopes_disabled():
-        p = o.payments.first()
-    assert p.provider == "free"
-    assert p.amount == o.total
-    assert p.state == "confirmed"
-    assert o.all_logentries().count() == 2
-
-
-@pytest.mark.django_db
-def test_order_create_invalid_payment_provider(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['payment_provider'] = 'foo'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'payment_provider': ['The given payment provider is not known.']}
-
-
-@pytest.mark.django_db
-def test_order_create_invalid_free_order(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['payment_provider'] = 'free'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == ['You cannot use the "free" payment provider for non-free orders.']
-
-
-@pytest.mark.django_db
-def test_order_create_invalid_status(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['status'] = 'e'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'status': ['"e" is not a valid choice.']}
-
-
-@pytest.mark.django_db
-def test_order_create_paid_generate_invoice(token_client, organizer, event, item, quota, question):
-    event.settings.invoice_generate = 'paid'
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['status'] = 'p'
-    res['payment_date'] = '2019-04-01 08:20:00Z'
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert o.invoices.count() == 1
-
-        p = o.payments.first()
-    assert p.provider == "banktransfer"
-    assert p.amount == o.total
-    assert p.state == "confirmed"
-    assert p.payment_date.year == 2019
-    assert p.payment_date.month == 4
-    assert p.payment_date.day == 1
-    assert p.payment_date.hour == 8
-    assert p.payment_date.minute == 20
-
-
-@pytest.fixture
-def seat(event, organizer, item):
-    SeatingPlan.objects.create(
-        name="Plan", organizer=organizer, layout="{}"
-    )
-    event.seat_category_mappings.create(
-        layout_category='Stalls', product=item
-    )
-    return event.seats.create(seat_number="A1", product=item, seat_guid="A1")
-
-
-@pytest.mark.django_db
-def test_order_create_with_seat(token_client, organizer, event, item, quota, seat, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['seat'] = seat.seat_guid
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.seat == seat
-
-
-@pytest.mark.django_db
-def test_order_create_with_blocked_seat_allowed(token_client, organizer, event, item, quota, seat, question):
-    seat.blocked = True
-    seat.save()
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['seat'] = seat.seat_guid
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['sales_channel'] = 'bar'
-    event.settings.seating_allow_blocked_seats_for_channel = ['bar']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-
-@pytest.mark.django_db
-def test_order_create_with_blocked_seat(token_client, organizer, event, item, quota, seat, question):
-    seat.blocked = True
-    seat.save()
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['seat'] = seat.seat_guid
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'seat': ['The selected seat "Seat A1" is not available.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_with_used_seat(token_client, organizer, event, item, quota, seat, question):
-    CartPosition.objects.create(
-        event=event, cart_id='aaa', item=item,
-        price=21.5, expires=now() + datetime.timedelta(minutes=10), seat=seat
-    )
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['seat'] = seat.seat_guid
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'seat': ['The selected seat "Seat A1" is not available.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_with_unknown_seat(token_client, organizer, event, item, quota, seat, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['seat'] = seat.seat_guid + '_'
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'seat': ['The specified seat does not exist.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_require_seat(token_client, organizer, event, item, quota, seat, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'seat': ['The specified product requires to choose a seat.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_unseated(token_client, organizer, event, item, quota, seat, question):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        quota.items.add(item2)
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item2.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['seat'] = seat.seat_guid
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'seat': ['The specified product does not allow to choose a seat.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_with_duplicate_seat(token_client, organizer, event, item, quota, seat, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'] = [
-        {
-            "positionid": 1,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": None,
-            "answers": [],
-            "subevent": None,
-            "seat": seat.seat_guid
-        },
-        {
-            "positionid": 2,
-            "item": item.pk,
-            "variation": None,
-            "price": "23.00",
-            "attendee_name_parts": {"full_name": "Peter"},
-            "attendee_email": None,
-            "addon_to": 1,
-            "answers": [],
-            "subevent": None,
-            "seat": seat.seat_guid
-        }
-    ]
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {},
-            {'seat': ['The selected seat "Seat A1" is not available.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_with_seat_consumed_from_cart(token_client, organizer, event, item, quota, seat, question):
-    CartPosition.objects.create(
-        event=event, cart_id='aaa', item=item,
-        price=21.5, expires=now() + datetime.timedelta(minutes=10), seat=seat
-    )
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['seat'] = seat.seat_guid
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['consume_carts'] = ['aaa']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.seat == seat
-
-
-@pytest.mark.django_db
-def test_order_create_send_no_emails(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 0
-
-
-@pytest.mark.django_db
-def test_order_create_send_emails(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['send_email'] = True
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 1
-    assert djmail.outbox[0].subject == "Your order: {}".format(resp.data['code'])
-
-
-@pytest.mark.django_db
-def test_order_create_send_emails_free(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['price'] = '0.00'
-    res['payment_provider'] = 'free'
-    del res['fees']
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['send_email'] = True
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 1
-    assert djmail.outbox[0].subject == "Your order: {}".format(resp.data['code'])
-
-
-@pytest.mark.django_db
-def test_order_create_send_emails_based_on_sales_channel(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['price'] = '0.00'
-    res['payment_provider'] = 'free'
-    del res['fees']
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['send_email'] = None
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 1
-    assert djmail.outbox[0].subject == "Your order: {}".format(resp.data['code'])
-
-    event.settingsmail_sales_channel_placed_paid = []
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 1
-
-
-@pytest.mark.django_db
-def test_order_create_send_emails_paid(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['send_email'] = True
-    res['status'] = 'p'
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 2
-    assert djmail.outbox[0].subject == "Your order: {}".format(resp.data['code'])
-    assert djmail.outbox[1].subject == "Payment received for your order: {}".format(resp.data['code'])
-
-
-@pytest.mark.django_db
-def test_order_create_send_emails_legacy(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['send_mail'] = True
-    res['status'] = 'p'
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    assert len(djmail.outbox) == 2
-    assert djmail.outbox[0].subject == "Your order: {}".format(resp.data['code'])
-    assert djmail.outbox[1].subject == "Payment received for your order: {}".format(resp.data['code'])
-
-
-@pytest.mark.django_db
-def test_order_paid_require_payment_method(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['payment_provider']
-    res['status'] = 'p'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == [
-        'You cannot create a paid order without a payment provider.'
-    ]
-
-    res['status'] = "n"
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        assert not o.payments.exists()
-
-
-@pytest.mark.django_db
-def test_order_create_auto_pricing(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['price']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.price == item.default_price
-    assert o.total == item.default_price + Decimal('0.25')
-
-
-@pytest.mark.django_db
-def test_order_create_auto_pricing_reverse_charge(token_client, organizer, event, item, quota, question, taxrule):
-    taxrule.eu_reverse_charge = True
-    taxrule.home_country = Country('DE')
-    taxrule.save()
-    item.tax_rule = taxrule
-    item.save()
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['invoice_address']['country'] = 'FR'
-    res['invoice_address']['is_business'] = True
-    res['invoice_address']['vat_id'] = 'FR12345'
-    res['invoice_address']['vat_id_validated'] = True
-    del res['positions'][0]['price']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.price == Decimal('19.33')
-    assert p.tax_rate == Decimal('0.00')
-    assert p.tax_value == Decimal('0.00')
-    assert o.total == Decimal('19.58')
-
-
-@pytest.mark.django_db
-def test_order_create_auto_pricing_country_rate(token_client, organizer, event, item, quota, question, taxrule):
-    taxrule.eu_reverse_charge = True
-    taxrule.custom_rules = json.dumps([
-        {'country': 'FR', 'address_type': '', 'action': 'vat', 'rate': '100.00'}
-    ])
-    taxrule.save()
-    item.tax_rule = taxrule
-    item.save()
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['invoice_address']['country'] = 'FR'
-    res['invoice_address']['is_business'] = True
-    res['invoice_address']['vat_id'] = 'FR12345'
-    res['invoice_address']['vat_id_validated'] = True
-    del res['positions'][0]['price']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.price == Decimal('38.66')
-    assert p.tax_rate == Decimal('100.00')
-    assert p.tax_value == Decimal('19.33')
-    assert o.total == Decimal('38.91')
-
-
-@pytest.mark.django_db
-def test_order_create_auto_pricing_reverse_charge_require_valid_vatid(token_client, organizer, event, item, quota,
-                                                                      question, taxrule):
-    taxrule.eu_reverse_charge = True
-    taxrule.home_country = Country('DE')
-    taxrule.save()
-    item.tax_rule = taxrule
-    item.save()
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['invoice_address']['country'] = 'FR'
-    res['invoice_address']['is_business'] = True
-    res['invoice_address']['vat_id'] = 'FR12345'
-    del res['positions'][0]['price']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.price == Decimal('23.00')
-    assert p.tax_rate == Decimal('19.00')
-
-
-@pytest.mark.django_db
-def test_order_create_autopricing_voucher_budget_partially(token_client, organizer, event, item, quota, question,
-                                                           taxrule):
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=21.50, item=item, budget=Decimal('2.50'),
-                                        max_usages=999)
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['voucher'] = voucher.code
-    del res['positions'][0]['price']
-    del res['positions'][0]['positionid']
-    res['positions'].append(res['positions'][0])
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-        p2 = o.positions.last()
-    assert p.price == Decimal('21.50')
-    assert p2.price == Decimal('22.00')
-
-
-@pytest.mark.django_db
-def test_order_create_autopricing_voucher_budget_full(token_client, organizer, event, item, quota, question, taxrule):
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=21.50, item=item, budget=Decimal('0.50'),
-                                        max_usages=999)
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['voucher'] = voucher.code
-    del res['positions'][0]['price']
-    del res['positions'][0]['positionid']
-    res['positions'].append(res['positions'][0])
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{}, {'voucher': ['The voucher has a remaining budget of 0.00, therefore a '
-                                                        'discount of 1.50 can not be given.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_budget_exceeded(token_client, organizer, event, item, quota, question, taxrule):
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=21.50, item=item, budget=Decimal('3.00'),
-                                        max_usages=999)
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['voucher'] = voucher.code
-    res['positions'][0]['price'] = '19.00'
-    del res['positions'][0]['positionid']
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'positions': [{'voucher': ['The voucher has a remaining budget of 3.00, therefore a '
-                                                    'discount of 4.00 can not be given.']}]}
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_price(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['price']
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=15, item=item)
-    res['positions'][0]['voucher'] = voucher.code
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        o = Order.objects.get(code=resp.data['code'])
-        p = o.positions.first()
-    assert p.voucher == voucher
-    voucher.refresh_from_db()
-    assert voucher.redeemed == 1
-    assert p.price == Decimal('15.00')
-    assert o.total == Decimal('15.25')
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_unknown_code(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['price']
-    with scopes_disabled():
-        event.vouchers.create(price_mode="set", value=15, item=item)
-    res['positions'][0]['voucher'] = "FOOBAR"
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'voucher': ['Object with code=FOOBAR does not exist.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_redeemed(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    del res['positions'][0]['price']
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=15, item=item, redeemed=1)
-    res['positions'][0]['voucher'] = voucher.code
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'voucher': ['The voucher has already been used the maximum number of times.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_redeemed_partially(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    res['positions'][0]['item'] = item.pk
-    del res['positions'][0]['price']
-    del res['positions'][0]['positionid']
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=15, item=item, redeemed=1, max_usages=2)
-    res['positions'][0]['voucher'] = voucher.code
-    res['positions'].append(copy.deepcopy(res['positions'][0]))
-    res['positions'].append(copy.deepcopy(res['positions'][0]))
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {},
-            {'voucher': ['The voucher has already been used the maximum number of times.']},
-            {'voucher': ['The voucher has already been used the maximum number of times.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_item_mismatch(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['price']
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        voucher = event.vouchers.create(price_mode="set", value=15, item=item2, redeemed=0)
-    res['positions'][0]['voucher'] = voucher.code
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'voucher': ['This voucher is not valid for this product.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_expired(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['price']
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=15, item=item, redeemed=0,
-                                        valid_until=now() - datetime.timedelta(days=1))
-    res['positions'][0]['voucher'] = voucher.code
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-    assert resp.data == {
-        'positions': [
-            {'voucher': ['This voucher is expired.']},
-        ]
-    }
-
-
-@pytest.mark.django_db
-def test_order_create_voucher_block_quota(token_client, organizer, event, item, quota, question):
-    res = copy.deepcopy(ORDER_CREATE_PAYLOAD)
-    res['positions'][0]['item'] = item.pk
-    res['positions'][0]['answers'][0]['question'] = question.pk
-    del res['positions'][0]['price']
-    quota.size = 0
-    quota.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 400
-
-    with scopes_disabled():
-        voucher = event.vouchers.create(price_mode="set", value=15, item=item, redeemed=0,
-                                        block_quota=True)
-    res['positions'][0]['voucher'] = voucher.code
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/'.format(
-            organizer.slug, event.slug
-        ), format='json', data=res
-    )
-    assert resp.status_code == 201
-
-
 REFUND_CREATE_PAYLOAD = {
     "state": "created",
     "provider": "manual",
@@ -4203,634 +1642,6 @@ def test_order_delete_test_mode_voucher_cancelled_order(token_client, organizer,
 
 
 @pytest.mark.django_db
-def test_order_update_ignore_fields(token_client, organizer, event, order):
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'status': 'c'
-        }
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    assert order.status == 'n'
-
-
-@pytest.mark.django_db
-def test_order_update_only_partial(token_client, organizer, event, order):
-    resp = token_client.put(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'status': 'c'
-        }
-    )
-    assert resp.status_code == 405
-
-
-@pytest.mark.django_db
-def test_order_update_state_validation(token_client, organizer, event, order):
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'invoice_address': {
-                "is_business": False,
-                "company": "This is my company name",
-                "name": "John Doe",
-                "name_parts": {},
-                "street": "",
-                "state": "",
-                "zipcode": "",
-                "city": "Paris",
-                "country": "NONEXISTANT",
-                "internal_reference": "",
-                "vat_id": "",
-            }
-        }
-    )
-    assert resp.status_code == 400
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'invoice_address': {
-                "is_business": False,
-                "company": "This is my company name",
-                "name": "John Doe",
-                "name_parts": {},
-                "street": "",
-                "state": "NONEXISTANT",
-                "zipcode": "",
-                "city": "Test",
-                "country": "AU",
-                "internal_reference": "",
-                "vat_id": "",
-            }
-        }
-    )
-    assert resp.status_code == 400
-
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'invoice_address': {
-                "is_business": False,
-                "company": "This is my company name",
-                "name": "John Doe",
-                "name_parts": {},
-                "street": "",
-                "state": "QLD",
-                "zipcode": "",
-                "city": "Test",
-                "country": "AU",
-                "internal_reference": "",
-                "vat_id": "",
-            }
-        }
-    )
-    assert resp.status_code == 200
-    order.invoice_address.refresh_from_db()
-    assert order.invoice_address.state == "QLD"
-    assert order.invoice_address.country == "AU"
-
-
-@pytest.mark.django_db
-def test_order_update_allowed_fields(token_client, organizer, event, order):
-    event.settings.locales = ['de', 'en']
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'comment': 'Here is a comment',
-            'custom_followup_at': '2021-06-12',
-            'checkin_attention': True,
-            'email': 'foo@bar.com',
-            'phone': '+4962219999',
-            'locale': 'de',
-            'invoice_address': {
-                "is_business": False,
-                "company": "This is my company name",
-                "name": "John Doe",
-                "name_parts": {},
-                "street": "",
-                "state": "",
-                "zipcode": "",
-                "city": "Paris",
-                "country": "FR",
-                "internal_reference": "",
-                "vat_id": "",
-            }
-        }
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    assert order.comment == 'Here is a comment'
-    assert order.custom_followup_at.isoformat() == '2021-06-12'
-    assert order.checkin_attention
-    assert order.email == 'foo@bar.com'
-    assert order.phone == '+4962219999'
-    assert order.locale == 'de'
-    assert order.invoice_address.company == "This is my company name"
-    assert order.invoice_address.name_cached == "John Doe"
-    assert order.invoice_address.name_parts == {'_legacy': 'John Doe'}
-    assert str(order.invoice_address.country) == "FR"
-    assert not order.invoice_address.vat_id_validated
-    assert order.invoice_address.city == "Paris"
-    with scopes_disabled():
-        assert order.all_logentries().get(action_type='pretix.event.order.comment')
-        assert order.all_logentries().get(action_type='pretix.event.order.custom_followup_at')
-        assert order.all_logentries().get(action_type='pretix.event.order.checkin_attention')
-        assert order.all_logentries().get(action_type='pretix.event.order.contact.changed')
-        assert order.all_logentries().get(action_type='pretix.event.order.phone.changed')
-        assert order.all_logentries().get(action_type='pretix.event.order.locale.changed')
-        assert order.all_logentries().get(action_type='pretix.event.order.modified')
-
-
-@pytest.mark.django_db
-def test_order_update_validated_vat_id(token_client, organizer, event, order):
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'invoice_address': {
-                "is_business": False,
-                "company": "This is my company name",
-                "name": "John Doe",
-                "name_parts": {},
-                "street": "",
-                "state": "",
-                "zipcode": "",
-                "city": "Paris",
-                "country": "FR",
-                "internal_reference": "",
-                "vat_id": "FR123",
-                "vat_id_validated": True
-            }
-        }
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    assert order.invoice_address.vat_id == "FR123"
-    assert order.invoice_address.vat_id_validated
-
-
-@pytest.mark.django_db
-def test_order_update_invoiceaddress_delete_create(token_client, organizer, event, order):
-    event.settings.locales = ['de', 'en']
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'invoice_address': None,
-        }
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    with pytest.raises(InvoiceAddress.DoesNotExist):
-        order.invoice_address
-
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'invoice_address': {
-                "is_business": False,
-                "company": "This is my company name",
-                "name": "",
-                "name_parts": {},
-                "street": "",
-                "state": "",
-                "zipcode": "",
-                "city": "Paris",
-                "country": "Fr",
-                "internal_reference": "",
-                "vat_id": "",
-            }
-        }
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    assert order.invoice_address.company == "This is my company name"
-    assert str(order.invoice_address.country) == "FR"
-    assert order.invoice_address.city == "Paris"
-
-
-@pytest.mark.django_db
-def test_order_update_email_to_none(token_client, organizer, event, order):
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'email': None,
-        }
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    assert order.email is None
-
-
-@pytest.mark.django_db
-def test_order_update_locale_to_invalid(token_client, organizer, event, order):
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orders/{}/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={
-            'locale': 'de',
-        }
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'locale': ['"de" is not a supported locale for this event.']}
-
-
-@pytest.mark.django_db
-def test_order_create_invoice(token_client, organizer, event, order):
-    event.settings.invoice_generate = 'True'
-
-    event.settings.invoice_generate_sales_channels = []
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/create_invoice/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.status_code == 400
-
-    event.settings.invoice_generate_sales_channels = ['web']
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/create_invoice/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        pos = order.positions.first()
-    assert json.loads(json.dumps(resp.data)) == {
-        'order': 'FOO',
-        'number': 'DUMMY-00001',
-        'is_cancellation': False,
-        "invoice_from_name": "",
-        "invoice_from": "",
-        "invoice_from_zipcode": "",
-        "invoice_from_city": "",
-        "invoice_from_country": None,
-        "invoice_from_tax_id": "",
-        "invoice_from_vat_id": "",
-        "invoice_to": "Sample company\nNew Zealand\nVAT-ID: DE123",
-        "invoice_to_company": "Sample company",
-        "invoice_to_name": "",
-        "invoice_to_street": "",
-        "invoice_to_zipcode": "",
-        "invoice_to_city": "",
-        "invoice_to_state": "",
-        "invoice_to_country": "NZ",
-        "invoice_to_vat_id": "DE123",
-        "invoice_to_beneficiary": "",
-        "custom_field": None,
-        'date': now().date().isoformat(),
-        'refers': None,
-        'locale': 'en',
-        'introductory_text': '',
-        'additional_text': '',
-        'payment_provider_text': '',
-        'footer_text': '',
-        'lines': [
-            {
-                'position': 1,
-                'description': 'Budget Ticket<br />Attendee: Peter',
-                'subevent': None,
-                'event_date_from': '2017-12-27T10:00:00Z',
-                'event_date_to': None,
-                'event_location': None,
-                'fee_type': None,
-                'fee_internal_type': None,
-                'attendee_name': 'Peter',
-                'item': pos.item_id,
-                'variation': None,
-                'gross_value': '23.00',
-                'tax_value': '0.00',
-                'tax_rate': '0.00',
-                'tax_name': ''
-            },
-            {
-                'position': 2,
-                'description': 'Payment fee',
-                'subevent': None,
-                'event_date_from': '2017-12-27T10:00:00Z',
-                'event_date_to': None,
-                'event_location': None,
-                'fee_type': "payment",
-                'fee_internal_type': None,
-                'attendee_name': None,
-                'item': None,
-                'variation': None,
-                'gross_value': '0.25',
-                'tax_value': '0.05',
-                'tax_rate': '19.00',
-                'tax_name': ''
-            }
-        ],
-        'foreign_currency_display': None,
-        'foreign_currency_rate': None,
-        'foreign_currency_rate_date': None,
-        'internal_reference': ''
-    }
-
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/create_invoice/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.data == {'detail': 'An invoice for this order already exists.'}
-    assert resp.status_code == 400
-
-    event.settings.invoice_generate = 'False'
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/create_invoice/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.status_code == 400
-    assert resp.data == {'detail': 'You cannot generate an invoice for this order.'}
-
-
-@pytest.mark.django_db
-def test_order_regenerate_secrets(token_client, organizer, event, order):
-    s = order.secret
-    with scopes_disabled():
-        ps = order.positions.first().secret
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/regenerate_secrets/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.status_code == 200
-    order.refresh_from_db()
-    assert s != order.secret
-    with scopes_disabled():
-        assert ps != order.positions.first().secret
-
-
-@pytest.mark.django_db
-def test_order_resend_link(token_client, organizer, event, order):
-    djmail.outbox = []
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/resend_link/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.status_code == 204
-    assert len(djmail.outbox) == 1
-
-    order.email = None
-    order.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/resend_link/'.format(
-            organizer.slug, event.slug, order.code
-        ), format='json', data={}
-    )
-    assert resp.status_code == 400
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation(token_client, organizer, event, order, item):
-    with scopes_disabled():
-        op = order.positions.first()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('23.00'),
-        'gross_formatted': '23.00',
-        'name': '',
-        'net': Decimal('23.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': None,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_item_with_tax(token_client, organizer, event, order, item, taxrule):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23, tax_rule=taxrule)
-        op = order.positions.first()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('23.00'),
-        'gross_formatted': '23.00',
-        'name': '',
-        'net': Decimal('19.33'),
-        'rate': Decimal('19.00'),
-        'tax_rule': taxrule.pk,
-        'tax': Decimal('3.67')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_item_with_variation(token_client, organizer, event, order):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        var = item2.variations.create(default_price=12, value="XS")
-        op = order.positions.first()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-            'variation': var.pk
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('12.00'),
-        'gross_formatted': '12.00',
-        'name': '',
-        'net': Decimal('12.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': None,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_subevent(token_client, organizer, event, order, subevent):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        op = order.positions.first()
-    op.subevent = subevent
-    op.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-            'subevent': subevent.pk
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('23.00'),
-        'gross_formatted': '23.00',
-        'name': '',
-        'net': Decimal('23.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': None,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_subevent_with_override(token_client, organizer, event, order, subevent):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        se2 = event.subevents.create(name="Foobar", date_from=datetime.datetime(2017, 12, 27, 10, 0, 0, tzinfo=UTC))
-        se2.subeventitem_set.create(item=item2, price=12)
-        op = order.positions.first()
-    op.subevent = subevent
-    op.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-            'subevent': se2.pk
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('12.00'),
-        'gross_formatted': '12.00',
-        'name': '',
-        'net': Decimal('12.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': None,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_voucher_matching(token_client, organizer, event, order, subevent, item):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        q = event.quotas.create(name="Quota")
-        q.items.add(item)
-        q.items.add(item2)
-        voucher = event.vouchers.create(price_mode="set", value=15, quota=q)
-        op = order.positions.first()
-    op.voucher = voucher
-    op.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('15.00'),
-        'gross_formatted': '15.00',
-        'name': '',
-        'net': Decimal('15.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': None,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_voucher_not_matching(token_client, organizer, event, order, subevent, item):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        q = event.quotas.create(name="Quota")
-        q.items.add(item)
-        voucher = event.vouchers.create(price_mode="set", value=15, quota=q)
-        op = order.positions.first()
-    op.voucher = voucher
-    op.save()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('23.00'),
-        'gross_formatted': '23.00',
-        'name': '',
-        'net': Decimal('23.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': None,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_net_price(token_client, organizer, event, order, subevent, item, taxrule):
-    taxrule.price_includes_tax = False
-    taxrule.save()
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=10, tax_rule=taxrule)
-        op = order.positions.first()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('11.90'),
-        'gross_formatted': '11.90',
-        'name': '',
-        'net': Decimal('10.00'),
-        'rate': Decimal('19.00'),
-        'tax_rule': taxrule.pk,
-        'tax': Decimal('1.90')
-    }
-
-
-@pytest.mark.django_db
-def test_orderposition_price_calculation_reverse_charge(token_client, organizer, event, order, subevent, item, taxrule):
-    taxrule.price_includes_tax = False
-    taxrule.eu_reverse_charge = True
-    taxrule.home_country = Country('DE')
-    taxrule.save()
-    order.invoice_address.is_business = True
-    order.invoice_address.vat_id = 'ATU1234567'
-    order.invoice_address.vat_id_validated = True
-    order.invoice_address.country = Country('AT')
-    order.invoice_address.save()
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=10, tax_rule=taxrule)
-        op = order.positions.first()
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/price_calc/'.format(organizer.slug, event.slug, op.pk),
-        data={
-            'item': item2.pk,
-        }
-    )
-    assert resp.status_code == 200
-    assert resp.data == {
-        'gross': Decimal('10.00'),
-        'gross_formatted': '10.00',
-        'name': '',
-        'net': Decimal('10.00'),
-        'rate': Decimal('0.00'),
-        'tax_rule': taxrule.pk,
-        'tax': Decimal('0.00')
-    }
-
-
-@pytest.mark.django_db
 def test_revoked_secret_list(token_client, organizer, event):
     r = event.revoked_secrets.create(secret="abcd")
     res = {
@@ -4843,928 +1654,3 @@ def test_revoked_secret_list(token_client, organizer, event):
     ))
     assert resp.status_code == 200
     assert [res] == resp.data['results']
-
-
-@pytest.mark.django_db
-def test_position_update_ignore_fields(token_client, organizer, event, order):
-    with scopes_disabled():
-        op = order.positions.first()
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data={
-            'tax_rate': '99.99'
-        }
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.tax_rate == Decimal('0.00')
-
-
-@pytest.mark.django_db
-def test_position_update_only_partial(token_client, organizer, event, order):
-    with scopes_disabled():
-        op = order.positions.first()
-    resp = token_client.put(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data={
-            'price': '99.99'
-        }
-    )
-    assert resp.status_code == 405
-
-
-@pytest.mark.django_db
-def test_position_update_info(token_client, organizer, event, order, question):
-    with scopes_disabled():
-        op = order.positions.first()
-        question.type = Question.TYPE_CHOICE_MULTIPLE
-        question.save()
-        opt = question.options.create(answer="L")
-    payload = {
-        'company': 'VILE',
-        'attendee_name_parts': {
-            'full_name': 'Max Mustermann'
-        },
-        'street': 'Sesame Street 21',
-        'zipcode': '99999',
-        'city': 'Springfield',
-        'country': 'US',
-        'state': 'CA',
-        'attendee_email': 'foo@example.org',
-        'answers': [
-            {
-                'question': question.pk,
-                'answer': 'ignored',
-                'options': [opt.pk]
-            }
-        ]
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    assert resp.data['answers'] == [
-        {
-            'question': question.pk,
-            'question_identifier': question.identifier,
-            'answer': 'L',
-            'options': [opt.pk],
-            'option_identifiers': [opt.identifier],
-        }
-    ]
-    op.refresh_from_db()
-    assert op.company == 'VILE'
-    assert op.attendee_name_cached == 'Max Mustermann'
-    assert op.attendee_name_parts == {
-        '_scheme': 'full',
-        'full_name': 'Max Mustermann'
-    }
-    with scopes_disabled():
-        assert op.answers.get().answer == 'L'
-        assert op.street == 'Sesame Street 21'
-        assert op.zipcode == '99999'
-        assert op.city == 'Springfield'
-        assert str(op.country) == 'US'
-        assert op.state == 'CA'
-        assert op.attendee_email == 'foo@example.org'
-        le = order.all_logentries().last()
-    assert le.action_type == 'pretix.event.order.modified'
-    assert le.parsed_data == {
-        'data': [
-            {
-                'position': op.pk,
-                'company': 'VILE',
-                'attendee_name_parts': {
-                    '_scheme': 'full',
-                    'full_name': 'Max Mustermann'
-                },
-                'street': 'Sesame Street 21',
-                'zipcode': '99999',
-                'city': 'Springfield',
-                'country': 'US',
-                'state': 'CA',
-                'attendee_email': 'foo@example.org',
-                f'question_{question.pk}': 'L'
-            }
-        ]
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        assert order.all_logentries().last().pk == le.pk
-
-
-@pytest.mark.django_db
-def test_position_update_legacy_name(token_client, organizer, event, order):
-    with scopes_disabled():
-        op = order.positions.first()
-    payload = {
-        'attendee_name': 'Max Mustermann',
-        'attendee_name_parts': {
-            '_legacy': 'maria'
-        },
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    payload = {
-        'attendee_name': 'Max Mustermann',
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.attendee_name_cached == 'Max Mustermann'
-    assert op.attendee_name_parts == {
-        '_legacy': 'Max Mustermann'
-    }
-    with scopes_disabled():
-        assert op.answers.count() == 1  # answer does not get deleted
-
-
-@pytest.mark.django_db
-def test_position_update_state_validation(token_client, organizer, event, order):
-    with scopes_disabled():
-        op = order.positions.first()
-    payload = {
-        'country': 'DE',
-        'state': 'BW'
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-
-
-@pytest.mark.django_db
-def test_position_update_question_handling(token_client, organizer, event, order, question):
-    with scopes_disabled():
-        op = order.positions.first()
-    payload = {
-        'answers': [
-            {
-                'question': question.pk,
-                'answer': 'FOOBAR',
-            },
-            {
-                'question': question.pk,
-                'answer': 'FOOBAR',
-            },
-        ]
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    payload = {
-        'answers': [
-            {
-                'question': question.pk,
-                'answer': 'FOOBAR',
-            },
-        ]
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        assert op.answers.count() == 1
-    payload = {
-        'answers': [
-        ]
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        assert op.answers.count() == 0
-
-    r = token_client.post(
-        '/api/v1/upload',
-        data={
-            'media_type': 'image/png',
-            'file': ContentFile('file.png', 'invalid png content')
-        },
-        format='upload',
-        HTTP_CONTENT_DISPOSITION='attachment; filename="file.png"',
-    )
-    assert r.status_code == 201
-    file_id_png = r.data['id']
-
-    payload = {
-        'answers': [
-            {
-                "question": question.id,
-                "answer": file_id_png
-            }
-        ]
-    }
-    question.type = Question.TYPE_FILE
-    question.save()
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        answ = op.answers.get()
-    assert answ.file
-    assert answ.answer.startswith("file://")
-
-    payload = {
-        'answers': [
-            {
-                "question": question.id,
-                "answer": "file:keep"
-            }
-        ]
-    }
-    question.type = Question.TYPE_FILE
-    question.save()
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        answ = op.answers.get()
-    assert answ.file
-    assert answ.answer.startswith("file://")
-
-
-@pytest.mark.django_db
-def test_position_update_change_item(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        quota.items.add(item2)
-        op = order.positions.first()
-    payload = {
-        'item': item2.pk,
-    }
-    assert op.item != item2
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.item == item2
-
-
-@pytest.mark.django_db
-def test_position_update_change_item_wrong_event(token_client, organizer, event, event2, order, quota):
-    with scopes_disabled():
-        item2 = event2.items.create(name="Budget Ticket", default_price=23)
-        quota.items.add(item2)
-        op = order.positions.first()
-    payload = {
-        'item': item2.pk,
-    }
-    assert op.item != item2
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'object does not exist.' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_item_no_quota(token_client, organizer, event, order):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        op = order.positions.first()
-    payload = {
-        'item': item2.pk,
-    }
-    assert op.item != item2
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'quota' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_item_variation(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        v = item2.variations.create(value="foo")
-        quota.items.add(item2)
-        quota.variations.add(v)
-        op = order.positions.first()
-    payload = {
-        'item': item2.pk,
-        'variation': v.pk,
-    }
-    assert op.item != item2
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.item == item2
-    assert op.variation == v
-
-
-@pytest.mark.django_db
-def test_position_update_change_item_variation_required(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        v = item2.variations.create(value="foo")
-        quota.items.add(item2)
-        quota.variations.add(v)
-        op = order.positions.first()
-    payload = {
-        'item': item2.pk,
-    }
-    assert op.item != item2
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'variation' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_item_variation_mismatch(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        v = item2.variations.create(value="foo")
-        item3 = event.items.create(name="Budget Ticket", default_price=23)
-        v3 = item3.variations.create(value="foo")
-        quota.items.add(item2)
-        quota.items.add(item3)
-        quota.variations.add(v)
-        quota.variations.add(v3)
-        op = order.positions.first()
-    payload = {
-        'item': item2.pk,
-        'variation': v3.pk,
-    }
-    assert op.item != item2
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'variation' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_subevent(token_client, organizer, event, order, quota, item, subevent):
-    with scopes_disabled():
-        se2 = event.subevents.create(name="Foobar", date_from=datetime.datetime(2017, 12, 27, 10, 0, 0, tzinfo=UTC))
-        q2 = se2.quotas.create(name="foo", size=1, event=event)
-        q2.items.add(item)
-        op = order.positions.first()
-        op.subevent = subevent
-        op.save()
-    payload = {
-        'subevent': se2.pk,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.subevent == se2
-
-
-@pytest.mark.django_db
-def test_position_update_change_subevent_quota_empty(token_client, organizer, event, order, quota, item, subevent):
-    with scopes_disabled():
-        se2 = event.subevents.create(name="Foobar", date_from=datetime.datetime(2017, 12, 27, 10, 0, 0, tzinfo=UTC))
-        q2 = se2.quotas.create(name="foo", size=0, event=event)
-        q2.items.add(item)
-        op = order.positions.first()
-        op.subevent = subevent
-        op.save()
-    payload = {
-        'subevent': se2.pk,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'quota' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_seat(token_client, organizer, event, order, quota, item, seat):
-    with scopes_disabled():
-        seat2 = event.seats.create(seat_number="A2", product=item, seat_guid="A2")
-        op = order.positions.first()
-        op.seat = seat
-        op.save()
-    payload = {
-        'seat': seat2.seat_guid,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.seat == seat2
-
-
-@pytest.mark.django_db
-def test_position_update_unset_seat(token_client, organizer, event, order, quota, item, seat):
-    with scopes_disabled():
-        op = order.positions.first()
-        op.seat = seat
-        op.save()
-    payload = {
-        'seat': None,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.seat is None
-
-
-@pytest.mark.django_db
-def test_position_update_change_seat_taken(token_client, organizer, event, order, quota, item, seat):
-    with scopes_disabled():
-        seat2 = event.seats.create(seat_number="A2", product=item, seat_guid="A2", blocked=True)
-        op = order.positions.first()
-        op.seat = seat
-        op.save()
-    payload = {
-        'seat': seat2.seat_guid,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'seat' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_subevent_keep_seat(token_client, organizer, event, order, quota, item, subevent, seat):
-    with scopes_disabled():
-        seat.subevent = subevent
-        seat.save()
-        se2 = event.subevents.create(name="Foobar", date_from=datetime.datetime(2017, 12, 27, 10, 0, 0, tzinfo=UTC))
-        seat2 = event.seats.create(seat_number="A1", product=item, seat_guid="A1", subevent=se2)
-        q2 = se2.quotas.create(name="foo", size=1, event=event)
-        q2.items.add(item)
-        op = order.positions.first()
-        op.subevent = subevent
-        op.seat = seat
-        op.save()
-    payload = {
-        'subevent': se2.pk,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.subevent == se2
-    assert op.seat == seat2
-
-
-@pytest.mark.django_db
-def test_position_update_change_subevent_missing_seat(token_client, organizer, event, order, quota, item, subevent, seat):
-    with scopes_disabled():
-        seat.subevent = subevent
-        seat.save()
-        se2 = event.subevents.create(name="Foobar", date_from=datetime.datetime(2017, 12, 27, 10, 0, 0, tzinfo=UTC))
-        q2 = se2.quotas.create(name="foo", size=1, event=event)
-        q2.items.add(item)
-        op = order.positions.first()
-        op.subevent = subevent
-        op.seat = seat
-        op.save()
-    payload = {
-        'subevent': se2.pk,
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'seat' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_update_change_price(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        op = order.positions.first()
-    payload = {
-        'price': Decimal('119.00')
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.price == Decimal('119.00')
-    assert op.tax_rate == Decimal('0.00')
-    assert op.tax_value == Decimal('0.00')
-
-
-@pytest.mark.django_db
-def test_position_update_change_price_and_tax_rule(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        op = order.positions.first()
-        tr = event.tax_rules.create(rate=19)
-    payload = {
-        'price': Decimal('119.00'),
-        'tax_rule': tr.pk
-    }
-    resp = token_client.patch(
-        '/api/v1/organizers/{}/events/{}/orderpositions/{}/'.format(
-            organizer.slug, event.slug, op.pk
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    op.refresh_from_db()
-    assert op.price == Decimal('119.00')
-    assert op.tax_rate == Decimal('19.00')
-    assert op.tax_value == Decimal('19.00')
-    assert op.tax_rule == tr
-
-
-@pytest.mark.django_db
-def test_position_add_simple(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert order.positions.count() == 2
-        op = order.positions.last()
-        assert op.item == item
-        assert op.price == item.default_price
-        assert op.positionid == 3
-
-
-@pytest.mark.django_db
-def test_position_add_price(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-        'price': '99.99'
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert order.positions.count() == 2
-        op = order.positions.last()
-        assert op.item == item
-        assert op.price == Decimal('99.99')
-        assert op.positionid == 3
-
-
-@pytest.mark.django_db
-def test_position_add_subevent(token_client, organizer, event, order, quota, item, subevent):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-        quota.subevent = subevent
-        quota.save()
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-        'subevent': subevent.pk,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert order.positions.count() == 2
-        op = order.positions.last()
-        assert op.item == item
-        assert op.price == item.default_price
-        assert op.positionid == 3
-        assert op.subevent == subevent
-
-
-@pytest.mark.django_db
-def test_position_add_subevent_required(token_client, organizer, event, order, quota, item, subevent):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'subevent' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_add_quota_empty(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-        quota.size = 1
-        quota.save()
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'quota' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_add_seat(token_client, organizer, event, order, quota, item, seat):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-        'seat': seat.seat_guid,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert order.positions.count() == 2
-        op = order.positions.last()
-        assert op.item == item
-        assert op.price == item.default_price
-        assert op.positionid == 3
-        assert op.seat == seat
-
-
-@pytest.mark.django_db
-def test_position_add_seat_required(token_client, organizer, event, order, quota, item, seat):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'seat' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_add_addon_to(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        cat = event.categories.create(name="Workshops")
-        item2 = event.items.create(name="WS1", default_price=23, category=cat)
-        quota.items.add(item2)
-        item.addons.create(addon_category=cat)
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item2.pk,
-        'addon_to': 1,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert order.positions.count() == 2
-        op = order.positions.last()
-        assert op.positionid == 3
-        assert op.addon_to.positionid == 1
-
-
-@pytest.mark.django_db
-def test_position_add_addon_to_canceled_position(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        cat = event.categories.create(name="Workshops")
-        item2 = event.items.create(name="WS1", default_price=23, category=cat)
-        quota.items.add(item2)
-        item.addons.create(addon_category=cat)
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item2.pk,
-        'addon_to': 2,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'unknown position' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_add_addon_to_wrong_product(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-        'addon_to': 1,
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 400
-    assert 'selected base position does not allow you to add this product as an add-on' in str(resp.data)
-
-
-@pytest.mark.django_db
-def test_position_add_and_set_info(token_client, organizer, event, order, question, item):
-    with scopes_disabled():
-        assert order.positions.count() == 1
-    payload = {
-        'order': order.code,
-        'item': item.pk,
-        'attendee_name': 'John Doe',
-        'answers': [
-            {
-                'question': question.pk,
-                'answer': 'FOOBAR',
-            },
-        ]
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orderpositions/'.format(
-            organizer.slug, event.slug,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 201
-    with scopes_disabled():
-        assert order.positions.count() == 2
-        op = order.positions.last()
-        assert op.item == item
-        assert op.price == item.default_price
-        assert op.positionid == 3
-        assert op.attendee_name == 'John Doe'
-        assert op.answers.count() == 1
-
-
-@pytest.mark.django_db
-def test_order_change_patch(token_client, organizer, event, order, quota):
-    with scopes_disabled():
-        item2 = event.items.create(name="Budget Ticket", default_price=23)
-        quota.items.add(item2)
-        p = order.positions.first()
-        f = order.fees.first()
-    payload = {
-        'patch_positions': [
-            {
-                'position': p.pk,
-                'body': {
-                    'item': item2.pk,
-                    'price': '99.44',
-                },
-            },
-        ],
-        'patch_fees': [
-            {
-                'fee': f.pk,
-                'body': {
-                    'value': '10.00',
-                }
-            }
-        ]
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/change/'.format(
-            organizer.slug, event.slug, order.code,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        p.refresh_from_db()
-        assert p.price == Decimal('99.44')
-        assert p.item == item2
-        f.refresh_from_db()
-        assert f.value == Decimal('10.00')
-
-
-@pytest.mark.django_db
-def test_order_change_cancel_and_create(token_client, organizer, event, order, quota, item):
-    with scopes_disabled():
-        p = order.positions.first()
-        f = order.fees.first()
-    payload = {
-        'cancel_positions': [
-            {
-                'position': p.pk,
-            },
-        ],
-        'create_positions': [
-            {
-                'item': item.pk,
-                'price': '99.99'
-            },
-        ],
-        'cancel_fees': [
-            {
-                'fee': f.pk,
-            }
-        ]
-    }
-    resp = token_client.post(
-        '/api/v1/organizers/{}/events/{}/orders/{}/change/'.format(
-            organizer.slug, event.slug, order.code,
-        ), format='json', data=payload
-    )
-    assert resp.status_code == 200
-    with scopes_disabled():
-        p.refresh_from_db()
-        assert p.canceled
-        p_new = order.positions.last()
-        assert p_new != p
-        assert p_new.item == item
-        assert p_new.price == Decimal('99.99')
-        f.refresh_from_db()
-        assert f.canceled
-
-# todo: send_email
-# todo: reissue_invoice
-# todo: recalculate_taxes
-# todo: selection of positions/fees (right order, not canceled)
-# todo: change payload validation
